chore: separate dataagent eval tooling
