+++ .gitignore
@@ -42,6 +42,8 @@ dolphinscheduler-service/*.egg-info/
 /.claude/skills/*
 !/.claude/skills/opendata-frontend-skill/
 !/.claude/skills/opendata-frontend-skill/SKILL.md
+/.claude/skills/arch-governance-assistant/
+/dataagent/.claude/skills/arch-governance-assistant/
 *.swp
 *.swo
 *~
@@ -75,6 +77,9 @@ artifacts/
 
 # Test outputs
 integration-test-results/
+/reports/
+/evals/
+arch-governance-core.jsonl
 
 # Additional Maven module
 backend-agent-api/target/
