+++ dataagent/dataagent-backend/core/agent_runtime.py
@@ -63,6 +63,10 @@ def _build_system_prompt(database_hint: str | None, skill_runtime: dict[str, Any
         "- 对上游 / 下游 / 血缘问题，`run_sql.py` 默认会拒绝首轮 `data_lineage` 类 SQL；只有 lineage 快照仍缺必要字段时，才允许显式带 `DATAAGENT_ALLOW_LINEAGE_SQL_FALLBACK=1` 追加补充查询。",
         "- 用户要看 DDL / SHOW CREATE TABLE 时，优先 `mcp__portal__portal_get_table_ddl`；无 MCP 时优先 `get_table_ddl.py`。",
         "- 已启用 Skills 只提供 OpenDataWorks 平台术语、平台表和数据中台通用规则；不要臆造租户业务术语、业务环境默认值或隐藏口径。",
+        "- 当问题命中某个领域型 Skill 时，优先按该 Skill 的本体、术语和工具口径执行；不要先退回 OpenDataWorks 通用元数据搜索，除非用户明确询问平台元数据或 Skill 指明需要这么做。",
+        "- 对需要数据、清单、影响面、归因或风险判断的问题，能执行真实只读查询时，不要只返回待执行 SQL；必须先拿到工具结果，再基于结果回答。",
+        "- 遇到空结果、权限不足、工具超时或接口失败时，要说明已验证的查询口径、失败原因和最小下一步；不要继续换表、换字段、换路径或重复试探。",
+        "- 首次有效结果后结束当前查询链路；只有结果缺少回答问题所必需的字段或口径时，才追加一次最小补充查询。",
         "- 不要自己发明部署绝对路径、脚本名或命令格式；路径和参数以 skill 文档为准。",
         "- 阅读深度、执行顺序、是否先追问以及何时收口，都以当前 skill 文档和真实工具结果为准；不要把某个 skill 的局部流程提升成全局规则。",
         "- 遇到关键信息不明确时，优先依据当前 skill 和工具结果确认；仍无法确认再做最小追问。只允许只读执行。",
