+++ deploy/README.md
@@ -159,16 +159,16 @@ Use this method for isolated environments without internet access. You will use 
 
 ## 3. DataAgent Online Evaluation
 
-离线包内置架构治理 50 条核心用例，部署完成并配置可用模型后，可手动运行在线评测。builtin 与 DeepEval 是两个并列评测引擎，均位于离线包根目录 `evals/`，且均独立于 DataAgent runtime。
+离线包内置 DataAgent 架构治理评测工具，部署完成并配置可用模型后，可手动运行在线评测。builtin 与 DeepEval 是两个并列评测引擎，均位于离线包根目录 `tools/dataagent-evals/`，且均独立于 DataAgent runtime。私有评测集不随 GitHub 或离线包内置，运行时必须通过 `--dataset` 或 `DATAAGENT_EVAL_DATASET` 指定。
 
 ```bash
 DATAAGENT_EVAL_JUDGE_BASE_URL=https://api.example.com \
 DATAAGENT_EVAL_JUDGE_TOKEN=... \
 DATAAGENT_EVAL_JUDGE_MODEL=claude-opus-4-6 \
-bash scripts/run-dataagent-evals.sh --base-url http://127.0.0.1:8900
+bash scripts/run-dataagent-evals.sh --base-url http://127.0.0.1:8900 --dataset /path/to/private-cases.jsonl
 ```
 
-默认评测集路径为 `evals/dataagent-arch-governance/arch-governance-core.jsonl`。评测集和评测脚本位于离线包根目录，不放入 `deploy/dataagent-runtime/`。
+评测集由部署人员手动放置并通过 `--dataset` 指定。评测脚本位于离线包根目录，不放入 `deploy/dataagent-runtime/`。
 
 builtin 评测镜像为 `opendataworks-dataagent-evals-builtin:<tag>`，由 `scripts/run-dataagent-evals.sh` 默认调用。若只做本地 dry-run，可设置 `DATAAGENT_BUILTIN_RUN_LOCAL=1`。
 
@@ -176,16 +176,16 @@ builtin 评测镜像为 `opendataworks-dataagent-evals-builtin:<tag>`，由 `scr
 
 ```bash
 # 只校验评测集与报告目录，不调用服务
-DATAAGENT_BUILTIN_RUN_LOCAL=1 bash scripts/run-dataagent-evals.sh --dry-run
+DATAAGENT_BUILTIN_RUN_LOCAL=1 bash scripts/run-dataagent-evals.sh --dry-run --dataset /path/to/private-cases.jsonl
 
 # 只跑指定用例
-bash scripts/run-dataagent-evals.sh --case ARCH_ASSET_001 --case ARCH_EDGE_006
+bash scripts/run-dataagent-evals.sh --dataset /path/to/private-cases.jsonl --case CASE_ID
 
 # 覆盖 DataAgent 执行模型，并独立配置 judge 模型
 DATAAGENT_EVAL_JUDGE_BASE_URL=https://api.example.com \
 DATAAGENT_EVAL_JUDGE_TOKEN=... \
 DATAAGENT_EVAL_JUDGE_MODEL=claude-opus-4-6 \
-bash scripts/run-dataagent-evals.sh --provider-id openrouter --model anthropic/claude-sonnet-4.5
+bash scripts/run-dataagent-evals.sh --dataset /path/to/private-cases.jsonl --provider-id openrouter --model anthropic/claude-sonnet-4.5
 ```
 
 输出目录默认为当前离线包目录下的 `reports/dataagent-evals/<timestamp>/`。Docker/Podman 模式会把包目录挂载为 `/workspace`，默认输出会持久化回宿主机包目录而不是容器临时文件系统。输出包含：
@@ -199,7 +199,7 @@ bash scripts/run-dataagent-evals.sh --provider-id openrouter --model anthropic/c
 
 ### DeepEval Parallel Evaluation
 
-离线包也包含 DeepEval 并行评测镜像 `opendataworks-dataagent-evals-deepeval:<tag>`，用于和 builtin 评测结果横向比对。两个评测镜像都只包含评测工具和评测集，不包含 DataAgent backend，也不会随 `load-package-and-start.sh` 默认启动。
+离线包也包含 DeepEval 并行评测镜像 `opendataworks-dataagent-evals-deepeval:<tag>`，用于和 builtin 评测结果横向比对。两个评测镜像都只包含评测工具，不包含私有评测集或 DataAgent backend，也不会随 `load-package-and-start.sh` 默认启动。
 
 运行示例：
 
@@ -207,10 +207,10 @@ bash scripts/run-dataagent-evals.sh --provider-id openrouter --model anthropic/c
 DATAAGENT_EVAL_JUDGE_BASE_URL=https://api.example.com \
 DATAAGENT_EVAL_JUDGE_TOKEN=... \
 DATAAGENT_EVAL_JUDGE_MODEL=claude-opus-4-6 \
-bash scripts/run-dataagent-deepeval-evals.sh --base-url http://127.0.0.1:8900
+bash scripts/run-dataagent-deepeval-evals.sh --base-url http://127.0.0.1:8900 --dataset /path/to/private-cases.jsonl
 ```
 
-DeepEval 评测默认读取 `evals/dataagent-arch-governance/arch-governance-core.jsonl`，输出同样包含 `cases.jsonl`、`summary.json`、`report.md` 和 `raw/<case_id>.json`。
+DeepEval 评测同样要求通过 `--dataset` 指定私有 JSONL，输出包含 `cases.jsonl`、`summary.json`、`report.md` 和 `raw/<case_id>.json`。
 
 ---
