+++ docs/plans/2026-05-12-dataagent-online-evaluation-plan.md
@@ -2,9 +2,9 @@
 
 > **For agentic workers:** REQUIRED SUB-SKILL: Use superpowers:subagent-driven-development (recommended) or superpowers:executing-plans to implement this plan task-by-task. Steps use checkbox (`- [ ]`) syntax for tracking.
 
-**Goal:** Add the first manually-triggered online DataAgent evaluation module for 50 architecture-governance cases.
+**Goal:** Add a manually-triggered online DataAgent evaluation module for externally supplied private architecture-governance cases.
 
-**Architecture:** Keep execution outside the main DataAgent runtime by adding an eval dataset under root `evals/` and a stdlib-only builtin runner under `evals/dataagent-arch-governance-builtin/`. The runner drives the real HTTP task chain and calls an independently configured judge endpoint; DataAgent backend remains only the system under test.
+**Architecture:** Keep execution outside the main DataAgent runtime by keeping private datasets out of GitHub and adding a stdlib-only builtin runner under `tools/dataagent-evals/builtin/`. The runner drives the real HTTP task chain and calls an independently configured judge endpoint; DataAgent backend remains only the system under test.
 
 **Tech Stack:** FastAPI, Pydantic, Claude Agent SDK, Python stdlib runner, Bash wrapper, JSONL/JSON/Markdown report artifacts.
 
@@ -12,13 +12,13 @@
 
 ## Files
 
-- Create: `evals/dataagent-arch-governance/arch-governance-core.jsonl`
+- Do not commit private datasets such as `arch-governance-core.jsonl`
 - Modify: `dataagent/dataagent-backend/main.py` only to ensure no eval router is registered
 - Modify: `dataagent/dataagent-backend/models/schemas.py` only to remove eval-only schemas
 - Create: `dataagent/dataagent-backend/tests/test_runtime_excludes_eval_api.py`
-- Create: `evals/dataagent-arch-governance-builtin/run.py`
-- Create: `evals/dataagent-arch-governance-builtin/Dockerfile`
-- Create: `evals/dataagent-arch-governance-builtin/README.md`
+- Create: `tools/dataagent-evals/builtin/run.py`
+- Create: `tools/dataagent-evals/builtin/Dockerfile`
+- Create: `tools/dataagent-evals/builtin/README.md`
 - Create: `scripts/run-dataagent-evals.sh`
 - Create: `scripts/run-dataagent-evals.py` compatibility shim
 - Create: `tests/test_run_dataagent_evals.py`
@@ -26,12 +26,13 @@
 - Modify: `scripts/README.md`
 - Modify: `deploy/README.md`
 
-## Task 1: Dataset
+## Task 1: Private Dataset Boundary
 
-- [ ] Create `evals/dataagent-arch-governance/arch-governance-core.jsonl` with 50 JSONL records from the source evaluation document.
+- [ ] Do not commit `arch-governance-core.jsonl` or any private case file.
+- [ ] Require `--dataset` or `DATAAGENT_EVAL_DATASET` to point at a private JSONL file.
 - [ ] Keep fixed fields: `case_id`, `category`, `question`, `expected_intent`, `expected_ontology_objects`, `expected_relations`, `expected_sql_or_tool_behavior`, `expected_answer_points`, `scoring`, `veto_rules`, `max_wait_seconds`.
 - [ ] Keep extension fields on every record as arrays/strings: `required_sql_fragments`, `forbidden_sql_patterns`, `expected_tool_names`, `judge_guidance`.
-- [ ] Verify the dataset has 50 records and unique case IDs with the runner dry-run test.
+- [ ] Verify the supplied dataset has unique case IDs with the runner dry-run test.
 
 ## Task 2: Runtime Boundary Tests
 
@@ -51,9 +52,9 @@
 ## Task 4: Runner Tests
 
 - [ ] Add `tests/test_run_dataagent_evals.py`.
-- [ ] Load `evals/dataagent-arch-governance-builtin/run.py` as a module through `importlib.util.spec_from_file_location`.
-- [ ] Test `--dry-run` validates the real dataset:
-  - 50 cases
+- [ ] Load `tools/dataagent-evals/builtin/run.py` as a module through `importlib.util.spec_from_file_location`.
+- [ ] Test `--dry-run` validates an external dataset:
+  - case count
   - unique `case_id`
   - scoring total equals 10
   - writes `summary.json` and `report.md`
@@ -79,7 +80,7 @@
 
 ## Task 5: Runner Implementation
 
-- [ ] Implement `evals/dataagent-arch-governance-builtin/run.py` with argparse options:
+- [ ] Implement `tools/dataagent-evals/builtin/run.py` with argparse options:
   - `--base-url`
   - `--dataset`
   - `--output-dir`
@@ -93,8 +94,7 @@
   - `--judge-model`
   - `--judge-timeout-seconds`
   - `--dry-run`
-- [ ] Implement dataset path resolution:
-  - prefer `evals/dataagent-arch-governance/arch-governance-core.jsonl`
+- [ ] Require an external dataset path and return exit code `2` when missing.
 - [ ] Implement sequential HTTP flow with `urllib.request`.
 - [ ] Extract evidence from task events and final messages:
   - tool names
@@ -109,7 +109,7 @@
 
 ## Task 6: Offline Packaging And Docs
 
-- [ ] Update `scripts/create-offline-package.sh` to copy root `evals/` into the package root and avoid copying eval assets into `deploy/dataagent-runtime/`.
+- [ ] Update `scripts/create-offline-package.sh` to copy `tools/dataagent-evals/` into the package root and avoid copying private datasets into `deploy/dataagent-runtime/`.
 - [ ] Add `opendataworks-dataagent-evals-builtin:<tag>` to image build, offline package, and load scripts.
 - [ ] Update `scripts/README.md` with online evaluation command examples.
 - [ ] Update `deploy/README.md` with offline package evaluation steps and artifact locations.
@@ -123,7 +123,7 @@
   - `python -m pytest tests/test_run_dataagent_evals.py -q`
 - [ ] Run packaging wrapper checks:
   - `bash scripts/run-dataagent-evals.sh --help`
-  - `python evals/dataagent-arch-governance-builtin/run.py --dry-run --dataset evals/dataagent-arch-governance/arch-governance-core.jsonl`
+  - `python tools/dataagent-evals/builtin/run.py --dry-run --dataset /path/to/private-cases.jsonl`
 - [ ] If a local full DataAgent environment is available, run:
-  - `bash scripts/run-dataagent-evals.sh --base-url http://127.0.0.1:8900 --case ARCH_ASSET_001 --case ARCH_EDGE_006`
+  - `bash scripts/run-dataagent-evals.sh --base-url http://127.0.0.1:8900 --dataset /path/to/private-cases.jsonl --case CASE_ID`
 - [ ] If the full online smoke is not run, report exactly which layers were verified and that real model/task-chain smoke remains untested.
