+++ docs/plans/2026-05-13-dataagent-deepeval-evaluation-plan.md
@@ -2,8 +2,8 @@
 
 ## Implementation Tasks
 
-- Add a standalone DeepEval runner under `evals/dataagent-arch-governance-deepeval/` with requirements, Dockerfile, and README.
-- Add the shared JSONL dataset under `evals/dataagent-arch-governance/` while leaving the existing builtin runner path intact.
+- Add a standalone DeepEval runner under `tools/dataagent-evals/deepeval/` with requirements, Dockerfile, and README.
+- Keep shared JSONL datasets private and external; both runners require `--dataset` or `DATAAGENT_EVAL_DATASET`.
 - Add `scripts/run-dataagent-deepeval-evals.sh` as the manual Docker/Podman entrypoint.
 - Add `opendataworks-dataagent-evals-deepeval:<tag>` to image build, offline package creation, and offline image loading scripts.
 - Update deploy and scripts documentation with DeepEval run commands and judge environment variables.
@@ -12,6 +12,6 @@
 
 - Unit test JSONL-to-DeepEval case conversion and custom metric normalization.
 - Contract test fake DataAgent HTTP flow plus fake DeepEval evaluation.
-- Contract test offline packaging scripts reference the DeepEval image and eval directory.
+- Contract test offline packaging scripts reference the DeepEval image and eval tool directory.
 - Run shell syntax checks for the touched scripts.
-- Run DeepEval runner dry-run against the shared dataset.
+- Run DeepEval runner dry-run against an external private dataset.
