+++ scripts/README.md
@@ -41,21 +41,21 @@ Run the stdlib-only architecture-governance evaluation suite through its Docker 
 DATAAGENT_EVAL_JUDGE_BASE_URL=https://api.example.com \
 DATAAGENT_EVAL_JUDGE_TOKEN=... \
 DATAAGENT_EVAL_JUDGE_MODEL=claude-opus-4-6 \
-bash scripts/run-dataagent-evals.sh --base-url http://127.0.0.1:8900
+bash scripts/run-dataagent-evals.sh --base-url http://127.0.0.1:8900 --dataset /path/to/private-cases.jsonl
 ```
 
 Useful options:
 ```bash
 # Local dry-run without Docker
-DATAAGENT_BUILTIN_RUN_LOCAL=1 bash scripts/run-dataagent-evals.sh --dry-run
+DATAAGENT_BUILTIN_RUN_LOCAL=1 bash scripts/run-dataagent-evals.sh --dry-run --dataset /path/to/private-cases.jsonl
 
 # Run selected cases only
-bash scripts/run-dataagent-evals.sh --case ARCH_ASSET_001 --case ARCH_EDGE_006
+bash scripts/run-dataagent-evals.sh --dataset /path/to/private-cases.jsonl --case CASE_ID
 ```
 
 Default output is written to `reports/dataagent-evals/<timestamp>/` with `cases.jsonl`, `summary.json`, `report.md`, and per-case raw JSON files. In Docker/Podman mode the wrapper mounts the package as `/workspace`, so the default output is persisted back to the host package directory rather than the temporary container filesystem.
 
-The builtin module lives at `evals/dataagent-arch-governance-builtin/` and uses image `opendataworks-dataagent-evals-builtin:<tag>`. It is outside the DataAgent runtime; DataAgent is only the system under test. The legacy `scripts/run-dataagent-evals.py` path remains as a compatibility shim for local Python execution.
+The builtin module lives at `tools/dataagent-evals/builtin/` and uses image `opendataworks-dataagent-evals-builtin:<tag>`. It is outside the DataAgent runtime; DataAgent is only the system under test. The legacy `scripts/run-dataagent-evals.py` path remains as a compatibility shim for local Python execution. Private datasets are not committed and must be supplied with `--dataset` or `DATAAGENT_EVAL_DATASET`.
 
 ### DataAgent DeepEval Evaluation
 Run the DeepEval-based parallel evaluation module through its Docker image:
@@ -63,16 +63,16 @@ Run the DeepEval-based parallel evaluation module through its Docker image:
 DATAAGENT_EVAL_JUDGE_BASE_URL=https://api.example.com \
 DATAAGENT_EVAL_JUDGE_TOKEN=... \
 DATAAGENT_EVAL_JUDGE_MODEL=claude-opus-4-6 \
-bash scripts/run-dataagent-deepeval-evals.sh --base-url http://127.0.0.1:8900
+bash scripts/run-dataagent-deepeval-evals.sh --base-url http://127.0.0.1:8900 --dataset /path/to/private-cases.jsonl
 ```
 
 Useful options mirror the builtin runner:
 ```bash
 # Local dry-run without Docker
-DATAAGENT_DEEPEVAL_RUN_LOCAL=1 bash scripts/run-dataagent-deepeval-evals.sh --dry-run
+DATAAGENT_DEEPEVAL_RUN_LOCAL=1 bash scripts/run-dataagent-deepeval-evals.sh --dry-run --dataset /path/to/private-cases.jsonl
 
 # Run selected cases only
-bash scripts/run-dataagent-deepeval-evals.sh --case ARCH_ASSET_001 --case ARCH_EDGE_006
+bash scripts/run-dataagent-deepeval-evals.sh --dataset /path/to/private-cases.jsonl --case CASE_ID
 ```
 
-The builtin and DeepEval runners use the shared dataset at `evals/dataagent-arch-governance/arch-governance-core.jsonl` and write the same report file layout.
+The builtin and DeepEval runners require an external private JSONL dataset and write the same report file layout.
