+++ tools/dataagent-evals/builtin/Dockerfile
@@ -1,0 +1,10 @@
+FROM python:3.11-slim
+
+ENV PYTHONDONTWRITEBYTECODE=1
+ENV PYTHONUNBUFFERED=1
+
+WORKDIR /app/tools/dataagent-evals/builtin
+
+COPY tools/dataagent-evals /app/tools/dataagent-evals
+
+ENTRYPOINT ["python", "/app/tools/dataagent-evals/builtin/run.py"]
