+++ tools/dataagent-evals/builtin/README.md
@@ -6,13 +6,7 @@ It is intentionally separate from the DataAgent backend runtime. DataAgent is on
 
 ## Dataset
 
-Default dataset:
-
-```bash
-evals/dataagent-arch-governance/arch-governance-core.jsonl
-```
-
-The same JSONL dataset is shared with the DeepEval runner.
+Datasets are private deployment assets and are not committed with this tool. Pass the JSONL file explicitly with `--dataset` or set `DATAAGENT_EVAL_DATASET`.
 
 ## Run With Docker
 
@@ -20,7 +14,7 @@ The same JSONL dataset is shared with the DeepEval runner.
 DATAAGENT_EVAL_JUDGE_BASE_URL=https://api.example.com \
 DATAAGENT_EVAL_JUDGE_TOKEN=... \
 DATAAGENT_EVAL_JUDGE_MODEL=claude-opus-4-6 \
-bash scripts/run-dataagent-evals.sh --base-url http://127.0.0.1:8900
+bash scripts/run-dataagent-evals.sh --base-url http://127.0.0.1:8900 --dataset /path/to/private-cases.jsonl
 ```
 
 The wrapper runs `opendataworks-dataagent-evals-builtin:latest` by default. Override with:
@@ -32,13 +26,13 @@ OPENDATAWORKS_DATAAGENT_EVALS_BUILTIN_IMAGE=opendataworks-dataagent-evals-builti
 ## Local Dry Run
 
 ```bash
-python3 evals/dataagent-arch-governance-builtin/run.py --dry-run
+python3 tools/dataagent-evals/builtin/run.py --dry-run --dataset /path/to/private-cases.jsonl
 ```
 
 or:
 
 ```bash
-DATAAGENT_BUILTIN_RUN_LOCAL=1 bash scripts/run-dataagent-evals.sh --dry-run
+DATAAGENT_BUILTIN_RUN_LOCAL=1 bash scripts/run-dataagent-evals.sh --dry-run --dataset /path/to/private-cases.jsonl
 ```
 
 Dry run validates the dataset and writes `summary.json` / `report.md` without calling DataAgent or the judge model.
